+++ src/services/tests/Demoulas.ProfitSharing.IntegrationTests/Reports/YearEnd/Update/PRFT.cs
@@ -4,7 +4,7 @@ public class PRFT
 {
     public long FD_BADGE { get; set; } // PIC 9(7).
     public string? FD_NAME { get; set; } // PIC X(25).
-    public long FD_SSN { get; set; } // PIC 9(9)
+    public int FD_SSN { get; set; } // PIC 9(9)
     public decimal FD_AMT { get; set; } // PIC S9(8)V99
     public decimal FD_DIST1 { get; set; } // PIC S9(8)V99
     public decimal FD_MIL { get; set; } // PIC S9(8)V99
